move evrything to a java package folder structure or whatever its called
