+++ src/commonMain/kotlin/io/github/detachhead/kotlinxpath/components/axis.kt
@@ -1,4 +1,4 @@
-package components
+package io.github.detachhead.kotlinxpath.components
 
 import kotlin.js.JsName
