+++ src/commonMain/kotlin/io/github/detachhead/kotlinxpath/internalHelpers/Strings.kt
